feat: enhance mobile tab bar and profile dialog
- Updated MobileTabBar.vue to include a new sidebar for profile management with improved styling and layout.
- Refactored ProfileDialog.vue to utilize ProfileContent component for better separation of concerns and cleaner code.
- Adjusted carousel store to remove profile card and update related logic for active cards.
- Enhanced styles for sidebar elements, buttons, and typography for a more cohesive design.
