+++ src/components/navbar/MobileTabBar.vue
@@ -22,19 +22,29 @@
 	<Sidebar
 		v-model:visible="showPreferences"
 		position="bottom"
-		class="h-auto rounded-t-2xl pb-4">
+		class="h-auto rounded-t-2xl pb-4 custom-mobile-sidebar"
+		:showCloseIcon="false">
 		<template #header>
-			<div class="flex items-center gap-2">
-				<i class="pi pi-cog text-xl"></i>
-				<span class="font-bold text-lg">App Preferences</span>
+			<div
+				class="flex items-center justify-between w-full pb-2 sidebar-header-border m-0">
+				<div class="flex items-center gap-2 sidebar-title-wrap">
+					<i class="pi pi-cog text-xl"></i>
+					<span class="sidebar-title">App Preferences</span>
+				</div>
+
+				<button
+					class="btn-close"
+					@click="showPreferences = false">
+					<i class="pi pi-times text-sm"></i>
+				</button>
 			</div>
 		</template>
 
-		<div class="flex flex-col gap-6 py-2 px-2">
+		<div class="flex flex-col gap-6 pt-4 pb-2 px-4">
 			<div class="flex items-center justify-between">
 				<div class="flex items-center gap-4">
 					<div
-						class="w-10 h-10 rounded-full bg-surface-100 dark:bg-surface-800 flex items-center justify-center">
+						class="w-10 h-10 rounded-full icon-circle flex items-center justify-center">
 						<i
 							:class="
 								preferencesStore.isDarkMode
@@ -43,8 +53,8 @@
 							"></i>
 					</div>
 					<div>
-						<h4 class="font-semibold m-0 text-c">Dark Mode</h4>
-						<p class="text-sm text-c m-0">Change app appearance</p>
+						<h4 class="sidebar-label">Dark Mode</h4>
+						<p class="sidebar-desc">Change app appearance</p>
 					</div>
 				</div>
 				<ToggleSwitch v-model="preferencesStore.isDarkMode" />
@@ -53,7 +63,7 @@
 			<div class="flex items-center justify-between">
 				<div class="flex items-center gap-4">
 					<div
-						class="w-10 h-10 rounded-full bg-surface-100 dark:bg-surface-800 flex items-center justify-center">
+						class="w-10 h-10 rounded-full icon-circle flex items-center justify-center">
 						<i
 							:class="
 								preferencesStore.soundEnabled
@@ -62,8 +72,8 @@
 							"></i>
 					</div>
 					<div>
-						<h4 class="font-semibold m-0 text-c">Sound Effects</h4>
-						<p class="text-sm text-c m-0">Play UI sounds</p>
+						<h4 class="sidebar-label">Sound Effects</h4>
+						<p class="sidebar-desc">Play UI sounds</p>
 					</div>
 				</div>
 				<ToggleSwitch v-model="preferencesStore.soundEnabled" />
@@ -72,7 +82,7 @@
 			<div class="flex items-center justify-between">
 				<div class="flex items-center gap-4">
 					<div
-						class="w-10 h-10 rounded-full bg-surface-100 dark:bg-surface-800 flex items-center justify-center">
+						class="w-10 h-10 rounded-full icon-circle flex items-center justify-center">
 						<i
 							:class="
 								preferencesStore.animationsEnabled
@@ -81,47 +91,43 @@
 							"></i>
 					</div>
 					<div>
-						<h4 class="font-semibold m-0 text-b">Animations</h4>
-						<p class="text-sm text-c m-0">Confetti and visual effects</p>
+						<h4 class="sidebar-label">Animations</h4>
+						<p class="sidebar-desc">Confetti and visual effects</p>
 					</div>
 				</div>
 				<ToggleSwitch v-model="preferencesStore.animationsEnabled" />
 			</div>
 
-			<div class="h-px w-full bg-surface-200 dark:bg-surface-700"></div>
+			<div class="sidebar-divider"></div>
 
 			<div class="flex items-center justify-between gap-4">
 				<div class="flex items-center gap-4">
-					<div
-						class="w-10 h-10 rounded-full bg-surface-100 dark:bg-surface-800 flex items-center justify-center">
+					<div class="icon-circle">
 						<i class="pi pi-shield text-orange-500"></i>
 					</div>
 					<div>
-						<h4 class="font-semibold m-0 text-c">Cookie Preferences</h4>
-						<p class="text-sm text-c m-0">Review and update consent</p>
+						<h4 class="sidebar-label">Cookie Preferences</h4>
+						<p class="sidebar-desc">Review and update consent</p>
 					</div>
 				</div>
 				<Button
 					label="Manage"
-					severity="warn"
 					size="small"
-					class="rounded-lg"
+					class="btn-primary"
 					@pointerup.stop="triggerManageCookies"
-					@touchend.stop.prevent="triggerManageCookies"
 					@click.stop.prevent="triggerManageCookies" />
 			</div>
 
-			<div class="h-px w-full bg-surface-200 dark:bg-surface-700"></div>
+			<div class="sidebar-divider"></div>
 
 			<div class="flex flex-col gap-3">
 				<div class="flex items-center gap-4">
-					<div
-						class="w-10 h-10 rounded-full bg-surface-100 dark:bg-surface-800 flex items-center justify-center">
-						<i class="pi pi-file-edit text-orange-500"></i>
+					<div class="w-10 h-10 rounded-full flex items-center justify-center">
+						<i class="pi pi-file-edit icon-circle"></i>
 					</div>
 					<div>
-						<h4 class="font-semibold m-0 text-c">Legal</h4>
-						<p class="text-sm text-c m-0">Privacy and service terms</p>
+						<h4 class="sidebar-label">Legal</h4>
+						<p class="sidebar-desc">Privacy and service terms</p>
 					</div>
 				</div>
 				<div class="grid w-full grid-cols-1 gap-2">
@@ -141,6 +147,31 @@
 			</div>
 		</div>
 	</Sidebar>
+	<Sidebar
+		v-model:visible="showProfile"
+		position="bottom"
+		class="h-auto max-h-[90vh] rounded-t-2xl pb-4 custom-mobile-sidebar"
+		:showCloseIcon="false">
+		<template #header>
+			<div
+				class="flex items-center justify-between w-full pb-2 sidebar-header-border m-0">
+				<div class="flex items-center gap-2 sidebar-title-wrap">
+					<i class="pi pi-user text-xl"></i>
+					<span class="sidebar-title">Account</span>
+				</div>
+
+				<button
+					class="btn-close"
+					@click="showProfile = false">
+					<i class="pi pi-times text-sm"></i>
+				</button>
+			</div>
+		</template>
+
+		<div class="px-4 pb-4 overflow-y-auto">
+			<ProfileContent :isVisible="showProfile" />
+		</div>
+	</Sidebar>
 
 	<LegalDocumentsDialog ref="legalDialogRef" />
 </template>
@@ -151,6 +182,7 @@ import { storeToRefs } from "pinia";
 import { useCarouselStore } from "@/stores/useCarouselStore";
 import LegalDocumentsDialog from "@/components/legal/LegalDocumentsDialog.vue";
 import { useAuthStore } from "@/stores/auth";
+import ProfileContent from "@/components/login_view/ProfileContent.vue";
 
 // 1. Zastępujemy commonStore naszym nowym preferencesStore
 import { usePreferencesStore } from "@/stores/userPreferences";
@@ -170,9 +202,10 @@ const legalDialogRef = ref<{
 } | null>(null);
 
 const showPreferences = ref(false); // Steruje wysuwaniem panelu z dołu
+const showProfile = ref(false); // Steruje wysuwaniem panelu profilu
 
 // Zakładki karuzelowe mapowane na ID karty
-const cardTabs = ["textAdd", "manage", "stats", "profile"] as const;
+const cardTabs = ["textAdd", "manage", "stats"] as const;
 
 const activeTab = computed(() => {
 	const id = carouselStore.activeCardId;
@@ -181,20 +214,20 @@ const activeTab = computed(() => {
 });
 
 const tabs = [
-	// 2. Zmieniamy Theme na Settings i ikonę na pi-cog
 	{ id: "settings", label: "Settings", icon: "pi-cog", type: "action" },
 	{ id: "textAdd", label: "To-do", icon: "pi-list-check", type: "card" },
 	{ id: "manage", label: "Habits", icon: "pi-star", type: "card" },
 	{ id: "stats", label: "Stats", icon: "pi-chart-bar", type: "card" },
-	{ id: "profile", label: "Profile", icon: "pi-user", type: "card" },
+	{ id: "profile", label: "Profile", icon: "pi-user", type: "action" },
 ] as const;
 
 function onTab(tab: (typeof tabs)[number]) {
-	if (tab.type === "card") {
-		carouselStore.setActiveCard(tab.id as any);
+	if (tab.id === "profile") {
+		showProfile.value = true;
 	} else if (tab.id === "settings") {
-		// 3. Po kliknięciu w zębatkę, otwieramy panel boczny z dołu
 		showPreferences.value = true;
+	} else if (tab.type === "card") {
+		carouselStore.setActiveCard(tab.id as any);
 	}
 }
 
@@ -308,11 +341,267 @@ function openLegalDocument(document: "privacy" | "terms") {
 	background: var(--p-orange-400);
 }
 
+/* ====== TYPOGRAFIA SIDEBARÓW (Krok 2) ====== */
+.sidebar-title-wrap {
+	color: var(--p-gray-800);
+}
+.sidebar-title {
+	font-family: "Lora", serif;
+	font-size: 1.125rem;
+	font-weight: 700;
+}
+:where(.my-app-dark, .my-app-dark *) .sidebar-title-wrap {
+	color: var(--p-gray-100);
+}
+
+.sidebar-label {
+	font-family: "Lora", serif;
+	font-size: 0.95rem;
+	font-weight: 600;
+	color: var(--p-gray-800);
+	margin: 0;
+}
+:where(.my-app-dark, .my-app-dark *) .sidebar-label {
+	color: var(--p-gray-200);
+}
+
+.sidebar-desc {
+	font-family: "Lora", serif;
+	font-size: 0.8rem;
+	color: var(--p-gray-500);
+	margin: 0;
+}
+:where(.my-app-dark, .my-app-dark *) .sidebar-desc {
+	color: var(--p-gray-400);
+}
+
+/* Subtelne linie podziału wewnątrz ustawień */
+.sidebar-divider {
+	width: 100%;
+	height: 0; /* Ważne: 0, żeby tło nie pogrubiało linii */
+	background: transparent !important; /* Blokujemy narzuty Tailwinda */
+	border: none;
+	border-top: 1px solid color-mix(in srgb, var(--p-gray-200) 80%, transparent);
+	margin: 0;
+}
+.sidebar-header-border {
+	border-bottom: 1px solid
+		color-mix(in srgb, var(--p-gray-200) 80%, transparent);
+}
+:where(.my-app-dark, .my-app-dark *) .sidebar-divider,
+:where(.my-app-dark, .my-app-dark *) .sidebar-header-border {
+	border-color: color-mix(in srgb, var(--p-gray-700) 60%, transparent);
+}
+
 :where(.my-app-dark, .my-app-dark *) .tab-btn.active::before {
 	background: var(--p-orange-500);
 }
 
 .legal-action-btn:deep(.p-button-label) {
 	white-space: nowrap;
+}
+
+:deep(.custom-mobile-sidebar.p-sidebar) {
+	background: var(--p-surface-0, white) !important;
+	border: none !important;
+}
+
+:where(.my-app-dark, .my-app-dark *) :deep(.custom-mobile-sidebar.p-sidebar) {
+	background: var(--p-gray-900, #111827) !important;
+	/* Delikatna górna ramka (opcjonalnie, oddziela sidebar od czarnego tła ekranu) */
+	border-top: 1px solid color-mix(in srgb, var(--p-gray-700) 60%, transparent) !important;
+}
+:deep(.p-sidebar-header) {
+	padding: 0 !important;
+	background: transparent !important;
+	border: none !important;
+	margin: 0 !important;
+	display: block !important; /* Pozwala nam kontrolować strukturę w środku */
+}
+
+/* ====== PRZYCISKI I IKONY (Krok 3) ====== */
+/* Okrągłe tła pod ikonami */
+.icon-circle {
+	width: 2.5rem;
+	height: 2.5rem;
+	border-radius: 50%;
+	background: var(--p-gray-100);
+	display: flex;
+	align-items: center;
+	border: 1px solid transparent;
+	justify-content: center;
+}
+:where(.my-app-dark, .my-app-dark *) .icon-circle {
+	background: var(--p-gray-800);
+	border-color: var(--p-gray-700);
+}
+
+/* Główny przycisk akcji (np. Manage) */
+.btn-primary {
+	background: var(--p-orange-500) !important;
+	color: white !important;
+	border: none !important;
+	border-radius: 0.5rem !important;
+	font-family: "Lora", serif !important;
+}
+.btn-primary:hover {
+	background: var(--p-orange-600) !important;
+}
+
+/* Przyciski pomocnicze (np. Legal) */
+.btn-secondary {
+	background: var(--p-gray-100) !important;
+	color: var(--p-gray-700) !important;
+	border: none !important;
+	border-radius: 0.5rem !important;
+	font-family: "Lora", serif !important;
+}
+:where(.my-app-dark, .my-app-dark *) .btn-secondary {
+	background: var(--p-gray-800) !important;
+	color: var(--p-gray-200) !important;
+}
+.btn-secondary:hover {
+	background: var(--p-gray-200) !important;
+}
+:where(.my-app-dark, .my-app-dark *) .btn-secondary:hover {
+	background: var(--p-gray-700) !important;
+}
+
+.btn-close {
+	width: 2rem;
+	height: 2rem;
+	border-radius: 50%;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	border: none;
+	cursor: pointer;
+	background: var(--p-gray-100);
+	color: var(--p-gray-500);
+	transition:
+		background 0.2s ease,
+		color 0.2s ease;
+}
+.btn-close:hover {
+	background: var(--p-gray-200);
+	color: var(--p-gray-700);
+}
+:where(.my-app-dark, .my-app-dark *) .btn-close {
+	background: var(--p-gray-800);
+	color: var(--p-gray-400);
+}
+:where(.my-app-dark, .my-app-dark *) .btn-close:hover {
+	background: var(--p-gray-700);
+	color: var(--p-gray-200);
+}
+
+/* ====== 2. OKRĄGŁE TŁA POD IKONAMI ====== */
+.icon-circle {
+	width: 2.5rem;
+	height: 2.5rem;
+	border-radius: 50%;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	background: var(--p-gray-100);
+	flex-shrink: 0;
+}
+:where(.my-app-dark, .my-app-dark *) .icon-circle {
+	background: var(--p-gray-800);
+}
+
+/* ====== 3. GŁÓWNY PRZYCISK AKCJI (np. Manage) ====== */
+.btn-primary {
+	background: var(--p-orange-500) !important;
+	color: white !important;
+	border: none !important;
+	border-radius: 0.5rem !important;
+	font-family: "Lora", serif !important;
+	font-weight: 600 !important;
+	transition:
+		background 0.2s ease,
+		transform 0.1s ease !important;
+}
+.btn-primary:hover {
+	background: var(--p-orange-600) !important;
+}
+:where(.my-app-dark, .my-app-dark *) .btn-primary {
+	background: var(--p-orange-600) !important;
+}
+:where(.my-app-dark, .my-app-dark *) .btn-primary:hover {
+	background: var(--p-orange-500) !important;
+}
+
+/* ====== 4. PRZYCISKI POMOCNICZE (np. Legal) ====== */
+.btn-secondary {
+	background: var(--p-gray-100) !important;
+	color: var(--p-gray-700) !important;
+	border: none !important;
+	border-radius: 0.5rem !important;
+	font-family: "Lora", serif !important;
+	font-weight: 500 !important;
+	transition:
+		background 0.2s ease,
+		color 0.2s ease !important;
+}
+.btn-secondary:hover {
+	background: var(--p-gray-200) !important;
+	color: var(--p-gray-900) !important;
+}
+:where(.my-app-dark, .my-app-dark *) .btn-secondary {
+	background: var(--p-gray-800) !important;
+	color: var(--p-gray-300) !important;
+}
+:where(.my-app-dark, .my-app-dark *) .btn-secondary:hover {
+	background: var(--p-gray-700) !important;
+	color: var(--p-gray-100) !important;
+}
+
+:deep(.p-toggleswitch .p-toggleswitch-slider) {
+	background: var(--p-gray-300) !important;
+	border: none !important;
+	transition:
+		background-color 0.2s ease,
+		box-shadow 0.2s ease !important;
+}
+:deep(.p-toggleswitch .p-toggleswitch-slider::before) {
+	background: white !important;
+}
+
+/* Stan niezaznaczony (Dark) - Tło (ścieżka) i gałka */
+:where(.my-app-dark, .my-app-dark *)
+	:deep(
+		.p-toggleswitch:not(.p-toggleswitch-checked):not([data-p-checked="true"])
+			.p-toggleswitch-slider
+	) {
+	background: var(--p-gray-700) !important;
+}
+:where(.my-app-dark, .my-app-dark *)
+	:deep(
+		.p-toggleswitch:not(.p-toggleswitch-checked):not([data-p-checked="true"])
+			.p-toggleswitch-slider::before
+	) {
+	background: var(--p-gray-400) !important;
+}
+
+/* Stan zaznaczony (Light & Dark) - Zawsze pomarańczowy */
+:deep(.p-toggleswitch.p-toggleswitch-checked .p-toggleswitch-slider),
+:deep(.p-toggleswitch[data-p-checked="true"] .p-toggleswitch-slider) {
+	background: var(--p-orange-500) !important;
+}
+:deep(.p-toggleswitch.p-toggleswitch-checked .p-toggleswitch-slider::before),
+:deep(.p-toggleswitch[data-p-checked="true"] .p-toggleswitch-slider::before) {
+	background: white !important;
+}
+
+/* Hover na włączonym przełączniku */
+:deep(.p-toggleswitch.p-toggleswitch-checked:hover .p-toggleswitch-slider),
+:deep(.p-toggleswitch[data-p-checked="true"]:hover .p-toggleswitch-slider) {
+	background: var(--p-orange-600) !important;
+}
+
+/* Focus ring (estetyczna poświata przy kliknięciu/tabowaniu) */
+:deep(.p-toggleswitch:focus-within .p-toggleswitch-slider) {
+	box-shadow: 0 0 0 3px color-mix(in srgb, var(--p-orange-500) 25%, transparent) !important;
 }
 </style>
