+++ src/stores/useCarouselStore.ts
@@ -7,15 +7,18 @@ export const useCarouselStore = defineStore("carouselStore", () => {
 		{ id: "textAdd", order: 1 },
 		{ id: "manage", order: 2 },
 		{ id: "stats", order: 3 },
-		{ id: "profile", order: 4 },
 	];
 
 	const activeIndex = ref(1); // domyślnie "manage" (środkowa)
 
 	const activeCardId = computed(() => cards[activeIndex.value].id);
 	const activeCard = computed(() => cards[activeIndex.value]);
-	const leftCard = computed(() => activeIndex.value > 0 ? cards[activeIndex.value - 1] : null);
-	const rightCard = computed(() => activeIndex.value < cards.length - 1 ? cards[activeIndex.value + 1] : null);
+	const leftCard = computed(() =>
+		activeIndex.value > 0 ? cards[activeIndex.value - 1] : null,
+	);
+	const rightCard = computed(() =>
+		activeIndex.value < cards.length - 1 ? cards[activeIndex.value + 1] : null,
+	);
 
 	function setActiveCard(targetId: CarouselCardConfig["id"]) {
 		const idx = cards.findIndex((c) => c.id === targetId);
